NO DRC ERRORS!
I rewired a few things - moved stuff around on the tuning circuit,
moved the FM power trace a touch. I also did a cute thing with the
power trace on the left side of the board, and turned the antenna LC
circuit to 45*, and redrew the antenna restrict so that it’s all
smooth.