Some significant fixes & changes
So i started by looking at the alignment of the board, speaker, pot &
antenna. it turns out these were all a bit wacky - all laid out at
different grid spacings and none of them consistent. the alignment of
all was off relative to one another, and the board itself was just at
some arbitrary location. so:

i moved the whole board so that it’s centered at 35mm x 35mm
moved the speaker so it’s centered at 35mm x 35mm
moved the antenna and pot so they’re centered at 35mm (Ydim) and are
each 25.5mm from centerline in the Xdim
moved the batteries so that they’re centered at 35mm (Xdim) and each
18.5mm from centerline in the Ydim

then i:
added a wire for the batteries on the left side. i’d like to modify
this but it’s a start.
added bRestrict circles for the four bolts for the speaker. these are
just over 5mm in diameter, so they’re just a little bigger than our
nuts will be. there should be NO TRACES IN THESE ZONES!!!

i did *NOT* remove all the traces in those bRestricts, though - i need
to do other stuff at the moment.

there MAY BE SOME OTHER CHANGES NECESSARY as a result of my
modifications. i wanted to get all the mechanical components 100% set
and locked, so that’s all i’m sure of at the minute.